# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XpressPro FX Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kyc:
    get:
      operationId: getKycStatus
      tags: [kyc]
      summary: Get current user KYC status
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmissionRequest"
      responses:
        "200":
          description: Submitted KYC record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [withdrawals]
      summary: List the current user's withdrawals
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
    post:
      operationId: requestWithdrawal
      tags: [withdrawals]
      summary: Request a new withdrawal (enters pending state)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "200":
          description: Pending withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, awaiting_gas_fee, approved, rejected, completed, cancelled, expired]
      responses:
        "200":
          description: List of withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{withdrawalId}/decision:
    post:
      operationId: decideWithdrawal
      tags: [admin]
      summary: Approve or reject a pending withdrawal
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalDecisionRequest"
      responses:
        "200":
          description: Updated withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/kyc/{userId}/decision:
    post:
      operationId: decideKyc
      tags: [admin]
      summary: Approve or reject a KYC submission
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycDecisionRequest"
      responses:
        "200":
          description: Updated KYC record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycRecord"
...
  /admin/banks/{bankId}/verification:
    post:
      operationId: setBankVerification
      tags: [admin]
      summary: Set the verified flag on a linked bank account
      parameters:
        - name: bankId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankVerificationRequest"
      responses:
        "200":
          description: Updated bank account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBankSummary"
...
  /admin/cards/{cardId}/decision:
    post:
      operationId: setCardDecision
      tags: [admin]
      summary: Approve or reject a card request
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardDecisionRequest"
      responses:
        "200":
          description: Updated card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCardSummary"
  /admin/promotions:
    get:
      operationId: getAdminPromotions
      tags: [admin]
      summary: List all promotions (active + inactive)
      responses:
        "200":
          description: List of promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createPromotion
      tags: [admin]
      summary: Create a new promotion / activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromotionRequest"
      responses:
        "201":
          description: Created promotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
  /admin/promotions/{promotionId}:
    patch:
      operationId: updatePromotion
      tags: [admin]
      summary: Update an existing promotion
      parameters:
        - name: promotionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromotionRequest"
      responses:
        "200":
          description: Updated promotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
    delete:
      operationId: deletePromotion
      tags: [admin]
      summary: Delete a promotion
      parameters:
        - name: promotionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
  /admin/billing:
    get:
      operationId: getAdminBilling
      tags: [admin, billing]
      summary: Per-user billing overview for admin
      responses:
        "200":
          description: Admin billing rows + default rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBillingOverview"
    patch:
      operationId: updateBillingDefaults
      tags: [admin, billing]
      summary: Update platform default billing rates (applied to new users)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingRates"
      responses:
        "200":
          description: New defaults
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRates"
  /admin/billing/users/{userId}:
    patch:
      operationId: updateUserBillingRates
      tags: [admin, billing]
      summary: Set per-user billing rates (overrides defaults)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingRates"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBillingUserRow"
  /admin/billing/users/{userId}/mark-paid:
    post:
      operationId: adminMarkBillingPaid
      tags: [admin, billing]
      summary: Manually mark a user's current cycle items as paid (e.g. offline settlement)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingPaymentRequest"
      responses:
        "200":
          description: Updated row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBillingUserRow"
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    KycRecord:
      type: object
      properties:
        userId:
          type: string
        status:
          type: string
          enum: [not_submitted, pending, approved, rejected]
        idType:
          type: string
          nullable: true
        idNumber:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        submittedAt:
          type: string
          nullable: true
        decidedAt:
          type: string
          nullable: true
      required: [userId, status]
    KycSubmissionRequest:
      type: object
      properties:
        idType:
          type: string
          enum: [passport, drivers_license, national_id]
        idNumber:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        country:
          type: string
      required: [idType, idNumber, addressLine1, city, country]
    KycDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reason:
          type: string
          nullable: true
      required: [decision]
    Withdrawal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer]
        destination:
          type: string
        status:
          type: string
          enum: [pending, awaiting_gas_fee, approved, rejected, completed, cancelled, expired]
        rejectionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        decidedAt:
          type: string
          nullable: true
        gasFeeAmount:
          type: number
          nullable: true
          description: ETH gas fee amount the admin set on this withdrawal that the user must fund.
        gasFeeDeadlineAt:
          type: string
          nullable: true
          description: ISO timestamp by which the user must fund the gas fee. After this, the withdrawal expires.
        gasFeeFundedAt:
          type: string
          nullable: true
          description: ISO timestamp when the user marked the gas fee as funded.
        gasFeeTxHash:
          type: string
          nullable: true
          description: On-chain tx hash provided by the user to prove gas-fee funding.
        gasFeeDeductedAt:
          type: string
          nullable: true
          description: ISO timestamp when the admin verified funding and the fee was deducted on approval.
      required: [id, userId, userName, amount, currency, method, destination, status, createdAt]
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.000001
        currency:
          type: string
        method:
          type: string
          enum: [crypto_wallet, bank_transfer]
        destination:
          type: string
        sourceWalletId:
          type: string
          nullable: true
          description: |
            When supplied, the withdrawal is funded by directly broadcasting
            an on-chain transfer from this connected external wallet to the
            destination address. The server signs and broadcasts via
            ethers.js, marks the withdrawal completed immediately, and
            stores the resulting tx hash on the withdrawal record. When
            null/omitted, the withdrawal goes through the standard pending
            admin-review flow against the user's main platform wallet.
      required: [amount, currency, method, destination]
    WithdrawalDecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reason:
          type: string
          nullable: true
      required: [decision]
...
    AdminBankSummary:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userEmail:
          type: string
        bankName:
          type: string
        accountHolder:
          type: string
        last4:
          type: string
        currency:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, userName, userEmail, bankName, accountHolder, last4, currency, verified, createdAt]
    BankVerificationRequest:
      type: object
      properties:
        verified:
          type: boolean
      required: [verified]
...